update code follow the PR's comment
