--- src/common/__test__/utils.test.ts
@@ -1,9 +1,0 @@
-import PastelUtils from '../utils/utils'
-
-describe('feature/common/utils', () => {
-  test('Remove whitespace of a string in utils/removeAllBreakChar', async () => {
-    const result = PastelUtils.removeAllBreakChar(' p-secret-extended-key ')
-
-    expect(result).toEqual('p-secret-extended-key')
-  })
-})
