+++ src/common/components/SelectAmount/index.ts
@@ -1,5 +1,8 @@
 import SelectAmount from './SelectAmount'
+import { generateStep } from './SelectAmount.utils'
 
 export type { TOption } from './SelectAmount'
+
+export { generateStep }
 
 export default SelectAmount
