+++ src/common/utils/PastelPromoCode.ts
@@ -1,6 +1,8 @@
 import fs from 'fs'
 import path from 'path'
 import dayjs from 'dayjs'
+import { toast } from 'react-toastify'
+import log from 'electron-log'
 import { WalletRPC } from 'api/pastel-rpc'
 import store from '../../redux/store'
 
@@ -9,79 +11,79 @@ export type TPastelPromoCode = {
   address: string
 }
 
-export default class PastelPromoCode {
-  static async getFileName(): Promise<string | null> {
-    const dir = store.getState().appInfo.pastelWalletDirPath
-    if (!dir) {
-      throw new Error("Can't get path of pastel promo code store")
-    }
-
-    if (!fs.existsSync(dir)) {
-      await fs.promises.mkdir(dir)
-    }
-
-    const fileName = path.join(dir, 'PastelPromoCode.json')
-    if (!fs.existsSync(fileName)) {
-      fs.createWriteStream(fileName)
-    }
-
-    return fileName
+export const getFileName = async (): Promise<string | null> => {
+  const dir = store.getState().appInfo.pastelWalletDirPath
+  if (!dir) {
+    throw new Error("Can't get path of pastel promo code store")
   }
 
-  static async writePastelPromoCode(
-    pastelPromoCode?: TPastelPromoCode[],
-  ): Promise<void> {
-    if (pastelPromoCode?.length) {
-      const fileName = await this.getFileName()
-
-      if (fileName) {
-        await fs.promises.writeFile(fileName, JSON.stringify(pastelPromoCode))
-      }
-    }
+  if (!fs.existsSync(dir)) {
+    await fs.promises.mkdir(dir)
   }
 
-  static async readPastelPromoCode(): Promise<TPastelPromoCode[]> {
-    const fileName = await this.getFileName()
+  const fileName = path.join(dir, 'PastelPromoCode.json')
+  fs.createWriteStream(fileName)
 
-    try {
-      if (fileName) {
-        const data = await fs.promises.readFile(fileName)
+  return fileName
+}
 
-        if (data.length) {
-          return JSON.parse(data.toString())
-        }
-      }
+export const writePastelPromoCode = async (
+  pastelPromoCode?: TPastelPromoCode[],
+): Promise<void> => {
+  if (pastelPromoCode?.length) {
+    const fileName = await getFileName()
 
-      return []
-    } catch (err) {
-      console.log(
-        `common/utils/PastelPromoCode readPastelPromoCode error: ${err}`,
-        err,
-      )
-      return []
+    if (fileName) {
+      await fs.promises.writeFile(fileName, JSON.stringify(pastelPromoCode))
     }
   }
+}
 
-  static async importPastelPromoCode(pastelPromoCode: string): Promise<string> {
-    try {
-      const walletRPC = new WalletRPC()
-      const address = await walletRPC.importPrivKey(pastelPromoCode, true)
-      if (address) {
-        const currentPromoCode = await PastelPromoCode.readPastelPromoCode()
-        const promoCode = currentPromoCode.filter(pc => pc.address === address)
+export const readPastelPromoCode = async (): Promise<TPastelPromoCode[]> => {
+  const fileName = await getFileName()
 
-        if (promoCode.length < 1) {
-          const newPromoCode = currentPromoCode.concat({
-            label: `Pastel Promo Code ${dayjs().format('MM/DD/YYYY-HH:mm')}`,
-            address,
-          })
-          PastelPromoCode.writePastelPromoCode(newPromoCode)
-        }
-        return address
+  try {
+    if (fileName) {
+      const data = await fs.promises.readFile(fileName)
+
+      if (data.length) {
+        return JSON.parse(data.toString())
       }
-      return ''
-    } catch {
-      return ''
     }
+
+    return []
+  } catch (err) {
+    log.error(
+      `common/utils/PastelPromoCode readPastelPromoCode error: ${err}`,
+      err,
+    )
+    return []
+  }
+}
+
+export const importPastelPromoCode = async (
+  pastelPromoCode: string,
+): Promise<string | null> => {
+  try {
+    const walletRPC = new WalletRPC()
+    const address = await walletRPC.importPrivKey(pastelPromoCode, true)
+    const currentPromoCodeList = await readPastelPromoCode()
+    const promoCode = currentPromoCodeList.filter(pc => pc.address === address)
+
+    if (promoCode.length < 1) {
+      const newPromoCode = currentPromoCodeList.concat({
+        label: `Pastel Promo Code ${dayjs().format('MM/DD/YYYY-HH:mm')}`,
+        address,
+      })
+      await writePastelPromoCode(newPromoCode)
+    }
+    return address
+  } catch (err) {
+    toast(err.message, { type: 'error' })
+    log.error(
+      `common/utils/PastelPromoCode importPastelPromoCode error: ${err.message}`,
+      err,
+    )
+    return null
   }
 }
