+++ src/features/onboarding/register/StepFee.tsx
@@ -1,6 +1,5 @@
 import React, { useState, FormEvent } from 'react'
 import cn from 'classnames'
-import { toast } from 'react-toastify'
 import { PaymentMethods, TCentralizedExchangeEntity } from './Regiser.state'
 import { PrevButton, NextButton } from './Buttons'
 import Radio from 'common/components/Radio/Radio'
@@ -9,11 +8,7 @@ import Link from 'common/components/Link'
 import styles from './Register.module.css'
 import { Clipboard } from 'common/components/Icons'
 import { useCurrencyName } from 'common/hooks/appInfo'
-import PastelUtils from 'common/utils/utils'
-import PastelPromoCode from 'common/utils/PastelPromoCode'
-import { WalletRPC } from 'api/pastel-rpc'
-import { formatNumber } from 'common/utils/format'
-import { createNewPastelID } from 'api/pastel-rpc/pastelid'
+import PromoCode from './PromoCode'
 
 export type TStepFeeProps = {
   paymentMethod: PaymentMethods
@@ -30,19 +25,10 @@ export type TStepFeeProps = {
   goBack(): void
 }
 
-const targetBalance = 1000
-
 const StepFee = (props: TStepFeeProps): JSX.Element => {
-  const walletRPC = new WalletRPC()
   const currencyName = useCurrencyName()
   const [copying, setCopying] = useState<boolean>(false)
   const [copied, setCopied] = useState<boolean>(false)
-  const [isValidPrivateKey, setValidPrivateKey] = useState<boolean>(false)
-  const [message, setMessage] = useState<string>('')
-  const [status, setStatus] = useState<string>('')
-  const [walletBalance, setWalletBalance] = useState(0)
-  const [promoBalance, setPromoBalance] = useState(0)
-  const [selectedAddress, seSelectedAddress] = useState<string>('')
 
   // maybe it would be better to load this list from somewhere
   const centralizedExs: TCentralizedExchangeEntity[] = [
@@ -76,49 +62,6 @@ const StepFee = (props: TStepFeeProps): JSX.Element => {
     setCopied(true)
   }
 
-  const handleNextClick = async (type?: string) => {
-    if (type === 'next') {
-      setStatus('loading')
-      try {
-        await createNewPastelID(props.password, selectedAddress)
-        props.finish()
-      } catch (error) {
-        toast.error(error.message)
-        setStatus('error')
-      }
-    } else {
-      setMessage('')
-      setValidPrivateKey(false)
-      if (PastelUtils.isValidPrivateKey(props.pastelPromoCode)) {
-        setStatus('loading')
-        const result = await PastelPromoCode.importPastelPromoCode(
-          props.pastelPromoCode,
-        )
-        if (result) {
-          const [addresses, totalBalances] = await Promise.all([
-            walletRPC.fetchTandZAddresses(),
-            walletRPC.fetchTotalBalance(),
-          ])
-          const promoCodeBalance = addresses.filter(
-            address => address.address === result,
-          )
-          setPromoBalance(
-            parseFloat(promoCodeBalance[0]?.amount.toString()) || 0,
-          )
-          setWalletBalance(totalBalances.total)
-          setStatus('done')
-          setValidPrivateKey(true)
-          seSelectedAddress(result)
-        } else {
-          setStatus('error')
-          setMessage('Promo Code is invalid.')
-        }
-      } else {
-        setMessage('Promo Code is invalid.')
-      }
-    }
-  }
-
   const showWarn =
     !copied &&
     ((props.paymentMethod === PaymentMethods.CentralizedExchange &&
@@ -134,197 +77,127 @@ const StepFee = (props: TStepFeeProps): JSX.Element => {
       props.centralizedExchangeName.length > 0 &&
       copied) ||
     (props.paymentMethod === PaymentMethods.DecentralizedExchange && copied) ||
-    props.paymentMethod === PaymentMethods.PslAddress ||
-    (props.paymentMethod === PaymentMethods.PastelPromoCode &&
-      props.pastelPromoCode.length > 0)
-
-  let promoCodeIsValid = null
-  if (!isValidPrivateKey && message) {
-    promoCodeIsValid = false
-  }
+    props.paymentMethod === PaymentMethods.PslAddress
 
   return (
     <div className='flex flex-col h-full'>
-      <div className='flex-grow pt-28'>
-        {props.paymentMethod !== PaymentMethods.AirdropPromoCode &&
-          props.paymentMethod !== PaymentMethods.PastelPromoCode && (
-            <>
-              <h1 className='text-gray-4a text-h3 font-extrabold'>
-                Choose Exchange to Purchase {currencyName}
-              </h1>
-              <h2 className='text-gray-71 text-base font-normal '>
-                Copy address
-              </h2>
+      {props.paymentMethod === PaymentMethods.PastelPromoCode ? (
+        <PromoCode
+          paymentMethod={props.paymentMethod}
+          pastelPromoCode={props.pastelPromoCode}
+          setPastelPromoCode={props.setPastelPromoCode}
+          password={props.password}
+          finish={props.finish}
+          goBack={props.goBack}
+        />
+      ) : (
+        <>
+          <div className='flex-grow pt-28'>
+            {props.paymentMethod !== PaymentMethods.AirdropPromoCode && (
+              <>
+                <h1 className='text-gray-4a text-h3 font-extrabold'>
+                  Choose Exchange to Purchase {currencyName}
+                </h1>
+                <h2 className='text-gray-71 text-base font-normal '>
+                  Copy address
+                </h2>
 
-              <div className='mt-7'>
-                <div
-                  className={cn(
-                    'w-full flex justify-between items-center border rounded px-4 py-2',
-                    showWarn ? 'border-orange-63' : 'border-gray-200',
-                  )}
-                >
-                  <div className='mr-3 overflow-hidden overflow-ellipsis h-5'>
-                    {props.exchangeAddress}
-                  </div>
-                  <button type='button' onClick={toClipboard}>
-                    <Clipboard
-                      size={20}
-                      className={cn(
-                        'cursor-pointer transition-transform duration-100 ease-in-out text-gray-88',
-                        copying ? 'transform scale-150' : '',
-                      )}
-                    />
-                  </button>
-                </div>
-                <div className='text-sm font-medium text-orange-63 h-4'>
-                  {showWarn ? 'copy your address first' : ''}
-                </div>
-              </div>
-
-              {props.paymentMethod === PaymentMethods.CentralizedExchange && (
-                <div className='mt-5'>
-                  <div className='text-base text-gray-a0'>
-                    Choose platform and pay
-                  </div>
-
-                  {centralizedExs.map((item, i) => (
-                    <div className='mt-3' key={i}>
-                      <Radio
-                        key={i}
-                        checked={props.centralizedExchangeName === item.name}
-                        onChange={val => onChangePayPlatform(item.name, val)}
-                      >
-                        <span
-                          className={cn(
-                            props.centralizedExchangeName === item.name
-                              ? 'text-gray-4a font-black text-base'
-                              : 'text-gray-4a text-opacity-50 text-base font-medium',
-                          )}
-                        >
-                          {item.name}
-                        </span>
-                      </Radio>
+                <div className='mt-7'>
+                  <div
+                    className={cn(
+                      'w-full flex justify-between items-center border rounded px-4 py-2',
+                      showWarn ? 'border-orange-63' : 'border-gray-200',
+                    )}
+                  >
+                    <div className='mr-3 overflow-hidden overflow-ellipsis h-5'>
+                      {props.exchangeAddress}
                     </div>
-                  ))}
+                    <button type='button' onClick={toClipboard}>
+                      <Clipboard
+                        size={20}
+                        className={cn(
+                          'cursor-pointer transition-transform duration-100 ease-in-out text-gray-88',
+                          copying ? 'transform scale-150' : '',
+                        )}
+                      />
+                    </button>
+                  </div>
+                  <div className='text-sm font-medium text-orange-63 h-4'>
+                    {showWarn ? 'copy your address first' : ''}
+                  </div>
                 </div>
-              )}
-            </>
-          )}
 
-        {props.paymentMethod === PaymentMethods.AirdropPromoCode && (
-          <>
-            <h1 className='text-gray-23 text-xl font-black'>
-              Airdrop Promo Code
-            </h1>
-            <h2 className='text-gray-77 text-sm font-normal'>
-              To Request a code, please fill in the form{' '}
-              <Link to='#' className='text-blue-3f'>
-                here
-              </Link>
-            </h2>
+                {props.paymentMethod === PaymentMethods.CentralizedExchange && (
+                  <div className='mt-5'>
+                    <div className='text-base text-gray-a0'>
+                      Choose platform and pay
+                    </div>
 
-            <div className={cn('mt-4 airdrop', styles.airdrop)}>
-              <Input
-                className='w-full'
-                type='text'
-                placeholder='Paste your promo code here'
-                onChange={(e: FormEvent<HTMLInputElement>) =>
-                  props.setPromoCode(e.currentTarget.value)
-                }
-              />
-              <div></div>
-            </div>
-          </>
-        )}
+                    {centralizedExs.map((item, i) => (
+                      <div className='mt-3' key={i}>
+                        <Radio
+                          key={i}
+                          checked={props.centralizedExchangeName === item.name}
+                          onChange={val => onChangePayPlatform(item.name, val)}
+                        >
+                          <span
+                            className={cn(
+                              props.centralizedExchangeName === item.name
+                                ? 'text-gray-4a font-black text-base'
+                                : 'text-gray-4a text-opacity-50 text-base font-medium',
+                            )}
+                          >
+                            {item.name}
+                          </span>
+                        </Radio>
+                      </div>
+                    ))}
+                  </div>
+                )}
+              </>
+            )}
 
-        {props.paymentMethod === PaymentMethods.PastelPromoCode && (
-          <>
-            <h1 className='text-gray-23 text-xl font-black'>
-              Pastel Promo Code
-            </h1>
-            <div className={cn('mt-4 airdrop', styles.airdrop)}>
-              <Input
-                className='w-full'
-                type='text'
-                placeholder='Paste your promo code here'
-                onChange={(e: FormEvent<HTMLInputElement>) =>
-                  props.setPastelPromoCode(e.currentTarget.value.trim())
-                }
-                isValid={promoCodeIsValid}
-                errorMessage={
-                  !promoCodeIsValid && message
-                    ? message || 'Promo Code is invalid.'
-                    : null
-                }
-                hint
-                hintAsTooltip={false}
-                value={props.pastelPromoCode}
-              />
-            </div>
-            {status === 'done' ? (
-              <div className='mt-6 text-gray-71 text-base font-normal'>
-                Congratulations, your personalized promotional code has been
-                accepted! You now have {formatNumber(promoBalance)}{' '}
-                {currencyName} in your wallet.{' '}
-                <button
-                  type='button'
-                  className='link'
-                  onClick={() => {
-                    props.setPastelPromoCode('')
-                    setStatus('')
-                    setMessage('')
-                    setValidPrivateKey(false)
-                  }}
-                >
-                  Add new pastel promo code.
-                </button>
-              </div>
-            ) : null}
-          </>
-        )}
-      </div>
+            {props.paymentMethod === PaymentMethods.AirdropPromoCode && (
+              <>
+                <h1 className='text-gray-23 text-xl font-black'>
+                  Airdrop Promo Code
+                </h1>
+                <h2 className='text-gray-77 text-sm font-normal'>
+                  To Request a code, please fill in the form{' '}
+                  <Link to='#' className='text-blue-3f'>
+                    here
+                  </Link>
+                </h2>
 
-      <div className='mt-7 flex justify-between'>
-        <PrevButton onClick={() => props.goBack()} />
-        {props.paymentMethod === PaymentMethods.PastelPromoCode ? (
-          <NextButton
-            className='min-w-160px'
-            onClick={() =>
-              handleNextClick(
-                status === 'done' && walletBalance >= targetBalance
-                  ? 'next'
-                  : 'apply',
-              )
-            }
-            text={
-              props.paymentMethod === PaymentMethods.PastelPromoCode &&
-              (walletBalance < targetBalance || !status) &&
-              status !== 'done'
-                ? props.paymentMethod === PaymentMethods.PastelPromoCode &&
-                  status === 'loading'
-                  ? 'Applying'
-                  : 'Apply'
-                : `Proceed to 1,000 ${currencyName} Payment`
-            }
-            disabled={
-              !nextActive ||
-              (status === 'done' && walletBalance < targetBalance) ||
-              status === 'loading'
-            }
-          />
-        ) : (
-          <NextButton
-            className='min-w-160px'
-            onClick={() => props.finish()}
-            text={
-              props.paymentMethod === PaymentMethods.AirdropPromoCode
-                ? 'Apply'
-                : `Proceed to 1,000 ${currencyName} Payment`
-            }
-            disabled={!nextActive}
-          />
-        )}
-      </div>
+                <div className={cn('mt-4 airdrop', styles.airdrop)}>
+                  <Input
+                    className='w-full'
+                    type='text'
+                    placeholder='Paste your promo code here'
+                    onChange={(e: FormEvent<HTMLInputElement>) =>
+                      props.setPromoCode(e.currentTarget.value)
+                    }
+                  />
+                  <div></div>
+                </div>
+              </>
+            )}
+          </div>
+          <div className='mt-7 flex justify-between'>
+            <PrevButton onClick={() => props.goBack()} />
+            <NextButton
+              className='min-w-160px'
+              onClick={() => props.finish()}
+              text={
+                props.paymentMethod === PaymentMethods.AirdropPromoCode
+                  ? 'Apply'
+                  : `Proceed to 1,000 ${currencyName} Payment`
+              }
+              disabled={!nextActive}
+            />
+          </div>
+        </>
+      )}
     </div>
   )
 }
