+++ src/features/wallet/TransactionHistoryModal.tsx
@@ -23,7 +23,7 @@ import {
 import dayjs from 'dayjs'
 import { AddressForm } from './AddressForm'
 import { useAddressBook } from 'common/hooks'
-import PastelUtils from 'common/utils/utils'
+import { isSapling, isTransparent } from 'common/utils/wallet'
 
 export type TTransactionHistoryModalProps = {
   isOpen: boolean
@@ -69,9 +69,7 @@ const TransactionHistoryModal = ({
   const getFilterAddresses = (trans: TTransaction[], isSource: boolean) => {
     const filtered = trans
       .filter(t => {
-        return isSource
-          ? PastelUtils.isSapling(t.address)
-          : !PastelUtils.isSapling(t.address)
+        return isSource ? isSapling(t.address) : !isSapling(t.address)
       })
       .map(t => {
         return {
@@ -197,9 +195,7 @@ const TransactionHistoryModal = ({
       custom: (value: string | number, row?: TRow) => {
         return (
           <div className='ml-46px flex items-center'>
-            {!PastelUtils.isTransparent(row?.address)
-              ? 'Shielded'
-              : 'Transparent'}
+            {!isTransparent(row?.address) ? 'Shielded' : 'Transparent'}
 
             {Math.floor(Math.random() * 10) % 2 ? (
               <div className='inline-block'>
